Fix bug with load config. Handle NOAA-APT failed

+++ automation/receiver.py
@@ -7,9 +7,7 @@ import sched
 import sys
 import typing
 
-import yaml
-
-from utils import first, get_receiver_command, get_planner_command
+from utils import first, get_receiver_command, get_planner_command, open_config
 from submitobs import submit_observation
 
 RECEIVER_COMMAND = get_receiver_command()
@@ -71,8 +69,7 @@ def decode_apt(input_path: str, output_path):
     ])
     process.wait()
 
-with open("config.yml") as f:
-    config = yaml.safe_load(f)
+config = open_config()
 
 if __name__ == '__main__':
     _, name, los, *_ = sys.argv
@@ -97,6 +94,6 @@ if __name__ == '__main__':
     png_filename = "/tmp/%s_%s.png" % (name, los)
     decode_apt(wav_filename, png_filename)
     os.remove(wav_filename)
-
-    submit_observation(png_filename, name, str(now_datetime), str(now_datetime), str(los_datetime), "")
-    os.remove(png_filename)
+    if os.path.exists(png_filename):
+        submit_observation(png_filename, name, str(now_datetime), str(now_datetime), str(los_datetime), "")
+        os.remove(png_filename)
